Fix barren hints checking from 800ca2a2
The previous code added checked for location hints if the region had been hinted, but did not have the region hints check if a location was hinted. That would lead to a barren hinted area receiving a location hint depending on the hint placement order. This fixes this symmetry so that the order doesn't matter.