Add an option to randomize more equipment and item colors
Added a single toggle in the equipment color section which randomizes the color of a few extra items.
But in theory all colors should be able to be set separately through cosmetics plando.

I might add more colors or make some of them into individual settings in the future, not sure.