Use Game Mode to determine whether to check for Ganon BK conditions.
Rename z64_file_t.interface_flag to game_mode (based on decomp) and check that it is 0 to give Ganon BK. It is seemingly 1 on the title cutscene, 2 on file select, and 0 during normal gameplay. This has already been used for CAN_USE_DPAD and CAN_DRAW_TRIFORCE in the past.