Better Mask De-equip Hack
The previous, more simplistic hack had a few issues. If there was nothing available to use on a C-button, the mask would still be de-equipped. If you had a mask equipped to a C-button but didn't have that mask in the child trade slot, trying to equip that mask would result in it being de-equipped. This resolves those issues.