Generalize Small Key Ring Logic
Instead of using a specific string replace on specific items, items can now specify an "alias" tuple in the special dictionary. The first value of the tuple is the name of the item that is being aliased, and the second value is the number of the aliased item to add in logic when the item is collected.