Define Junk Items in ItemList.py
Junk items are now defined by adding a 'junk' entry to the item's `special` dictionary. The value given is the weight that junk item gets in the junk_pool. If the weight is 0, the item is not added to the junk_pool but still is added to remove_junk_items. If the weight is negative, the item is still classified as junk but will not be in either junk_pool or remove_junk_items.