Starting ice traps
This allows an ice trap to be placed on an effective starting item or
given as an explicit starting item via plando. Behaves correctly
regarding saving and loading. Since the number of pending ice traps is
now stored in the save context, it is no longer possible to avoid an ice
trap by savewarping.