Add added/modified tricks to hellmode preset
Note: Some of the 'modified' tricks weren't present and needed to be added to the preset, so there are likely more missing tricks that will need to be added later. I only added the tricks that I touched on in this PR, so somebody will still have to update the entire list later.