Force Path of Magic hint
Prevents hinting any of the default goals, and adds a fixed hint for Path of Magic. I've tested it, and seems to consistently generate one Path of Magic hint with two copies. Unclear whether it should have one or two copies, since the numbers vary by type in this hint distro, but that's easy to change.