DMC Logic Bug
The change I made to allow destroying that boulder from below so you could go back from upper to lower later, it wasn't enforcing is adult, so the logic was allowing child to make use of it. Technically the seed is still possible in this case but good luck with that. I made no attempt to test this fix, so you might want to make sure seeds at least generate.