Version Refactor
Reworks version checks to make it easier for branches to have their own individual version tracking. Version checks are performed using the branch URL specified, and now checks the supplementary version as well.

The version bytes injected into the ROM header at address 0x35 has been expanded to 5 bytes. In addition to the 3 base version bytes, it now has branch identifier for the fourth byte and the supplementary version for the fifth byte.