Cosmetic Plando: Music Groups, Favorites, Excludes
Adds `bgm_groups` to cosmetic logs containing the following:

`favorites`: A list of sequence names that will be prioritized and will always be included unless there are more favorite sequences than tracks to shuffle them to.

`exclude`: A list of sequences that will be excluded. When randomizing music during patch file generation, these could still be included unless the number of target tracks is reduced somehow (such as assigning one sequence to multiple tracks or disabling tracks).

`groups`: A dictionary containing a name and a list of sequences. Inside of `bgm`, you can now substitute the name of a group (preceeded by a `#`, i.e. `#group_name`) and a random sequence from the list will be chosen. Once a sequence is chosen, it is removed from the list and won't be chosen again. If all sequences on the list are exhausted, the list will be refilled and the process continues.

Also allows direct list assignments in music plando.